+++ backend/src/test/java/com/staccato/comment/service/CommentServiceTest.java
@@ -19,8 +19,8 @@ import com.staccato.exception.StaccatoException;
 import com.staccato.fixture.CommentRequestFixture;
 import com.staccato.fixture.Member.MemberFixture;
 import com.staccato.fixture.memory.MemoryFixture;
-import com.staccato.fixture.moment.CommentFixture;
-import com.staccato.fixture.moment.CommentUpdateRequestFixture;
+import com.staccato.fixture.comment.CommentFixture;
+import com.staccato.fixture.comment.CommentUpdateRequestFixture;
 import com.staccato.fixture.moment.MomentFixture;
 import com.staccato.member.domain.Member;
 import com.staccato.member.repository.MemberRepository;
